deploy float to d

+++ src/main/java/com/arquitectura/app/modules/condensadores/CondService.java
@@ -75,30 +75,30 @@ public class CondService {
 	}
 	
 	public double obtenerTsiMuro(int zona, String tipo_ed) {
-		float u = condRepository.findByZona(zona).getU_muro();
-		float rsi = condRepository.findByZona(zona).getRsi_muro();
+		double u = condRepository.findByZona(zona).getU_muro();
+		double rsi = condRepository.findByZona(zona).getRsi_muro();
 		int te = condRepository.findByZona(zona).getTe_muro();
-		float ti = edTiRepository.findByTipo(tipo_ed).getTi();
+		double ti = edTiRepository.findByTipo(tipo_ed).getTi();
 		
 		double tsi = ti-u*rsi*(ti-te);
 		return Math.round(tsi*100)/100;
 	}
 	
 	public double obtenerTsiTecho(int zona, String tipo_ed) {
-		float u = condRepository.findByZona(zona).getU_techo();
-		float rsi = condRepository.findByZona(zona).getRsi_techo();
+		double u = condRepository.findByZona(zona).getU_techo();
+		double rsi = condRepository.findByZona(zona).getRsi_techo();
 		int te = condRepository.findByZona(zona).getTe_techo();
-		float ti = edTiRepository.findByTipo(tipo_ed).getTi();
+		double ti = edTiRepository.findByTipo(tipo_ed).getTi();
 		
 		double tsi = ti-u*rsi*(ti-te);
 		return Math.round(tsi*100)/100;
 	}
 	
 	public double obtenerTsiPiso(int zona, String tipo_ed) {
-		float u = condRepository.findByZona(zona).getU_piso();
-		float rsi = condRepository.findByZona(zona).getRsi_piso();
+		double u = condRepository.findByZona(zona).getU_piso();
+		double rsi = condRepository.findByZona(zona).getRsi_piso();
 		int te = condRepository.findByZona(zona).getTe_piso();
-		float ti = edTiRepository.findByTipo(tipo_ed).getTi();
+		double ti = edTiRepository.findByTipo(tipo_ed).getTi();
 		
 		double tsi = ti-u*rsi*(ti-te);
 		return Math.round(tsi*100)/100;
